Extensions: move utility functions to a utility module
Makes the cog cleaner and makes the functions more accessible for other
modules.
